Renames getLisings -> getListings
