+++ src/jobs/utils/getListings.ts
